api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.2",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "langfuse>=3.4.0",
    "mcp>=1.14.0",
    "pocketbase-async>=0.12.0",
    "pydantic>=2.11.9",
    "pydantic-ai>=1.0.8",
    "pydantic-settings>=2.10.1",
    "regex>=2025.9.18",
    "starlette>=0.48.0",
    "tiktoken>=0.11.0",
    "uvicorn>=0.35.0",
    "PyPDF2>=3.0.1",
    "PyMuPDF>=1.26.4",
    "pymupdf4llm>=0.0.27",
    "spacy>=3.8.7",
    "en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "Pillow>=11.3.0",
    "stripe>=12.5.1",
    "meilisearch-python-sdk>=4.10.1",

]

[tool.hatch.build.targets.wheel]
packages = ["src"]
